refactor: CurvedProgressBar를 정산 상세 화면으로 이동
