--- src/pages/expenseDetail/ui/CurvedProgressBar/index.style.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const Wrapper = styled.div`
-  margin-top: ${({ theme }) => theme.unit[20]};
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
