+++ src/pages/expenseDetail/ui/CurvedProgressBar/index.ts
